Moved init function to sperate file.
